restructure directory and remove unnecessary functions
