+++ others/tests/test_util_str.py
@@ -4,17 +4,7 @@ Test for utility.
 This module provides a test for utility.
 """
 
-import numpy as np
-from qtdraw.util.util_str import (
-    str_to_list,
-    is_regular_list,
-    apply_to_list,
-    apply_to_numpy,
-    to_fraction,
-    str_to_sympy,
-    get_variable,
-    affine_trans,
-)
+from qtdraw.sandbox.util import str_to_list, affine_trans
 
 # ==================================================
 # test
@@ -42,47 +32,6 @@ def test_str_to_list():
         print(str_to_list(i))
     for i in test2:
         print(str_to_list(i))
-
-
-# ==================================================
-def test_is_regular_list():
-    print("=== test_is_regular_list ===")
-    for i in test1:
-        print(is_regular_list(str_to_list(i)))
-    for i in test2:
-        print(is_regular_list(str_to_list(i)))
-
-
-# ==================================================
-def test_apply_to_list():
-    print("=== test_apply_to_list ===")
-    for i in test1:
-        print(apply_to_list(lambda x: len(x), str_to_list(i)))
-    for i in test2:
-        print(apply_to_list(lambda x: len(x), str_to_list(i)))
-    for i in test1:
-        print(apply_to_numpy(lambda x: len(x), np.array(str_to_list(i))))
-
-
-# ==================================================
-def test_to_fraction():
-    print("=== test_to_fraction ===")
-    data = [0.25, 0.3333333333333333, np.pi, np.e]
-    for i in data:
-        print(to_fraction(i))
-
-
-# ==================================================
-def test_str_to_sympy():
-    print("=== test_str_to_sympy ===")
-    test = ["x+y", "a", "1.3", "2 sqrt(3)/3", "sin(x)", "0.5cos(y)", "3+x"]
-    var = ["x", "y", "z"]
-    for i in test:
-        v = get_variable(str_to_sympy(i))
-        s1 = str_to_sympy(i, var, rational=False)
-        s2 = str_to_sympy(i, var, rational=True)
-        s3 = str_to_sympy(i, var, subs={"x": 0.1})
-        print(f"{i} {v} => {s1}, {s2}, {s3}")
 
 
 # ==================================================
@@ -121,8 +70,4 @@ def test_affine_trans():
 
 # ==================================================
 test_str_to_list()
-test_is_regular_list()
-test_apply_to_list()
-test_to_fraction()
-test_str_to_sympy()
 test_affine_trans()
