+++ qtdraw/multipie/util_multipie.py
@@ -6,7 +6,7 @@ This module contains utilities for multipie plugin.
 
 import sympy as sp
 from gcoreutils.nsarray import NSArray
-from qtdraw.util.util_str import str_to_list
+from qtdraw.sandbox.util import str_to_list
 
 
 # ==================================================
